fix: harden auth logging lifecycle and stats overflow handling
Agent-Logs-Url: https://github.com/Majd868/Hisham-s-project-/sessions/a8761a6e-8105-4c7e-a971-0f7498e758a4

+++ app/src/main/java/com/majd/pomodoro/HomeActivity.java
@@ -122,7 +122,7 @@ public class HomeActivity extends AppCompatActivity {
     private void onSessionCompleted() {
         stopPomodoro();
         statusText.setText(R.string.status_completed);
-        int focusMinutes = engine.getFocusMin() * engine.getTotalBlocks();
+        long focusMinutes = (long) engine.getFocusMin() * engine.getTotalBlocks();
         prefsManager.trackCompletedSession(focusMinutes);
         repository.saveStudyState(
                 prefsManager.getFocusMin(),
